VLM-Bot/requirements.txt
# ============================================================================
# VLM-Bot Requirements - Python 3.10+
# Compatible avec: NVIDIA RTX 3050 Laptop (4GB VRAM, CUDA 13.1)
# ============================================================================

# ⚠️ IMPORTANT: Installer PyTorch AVANT avec conda:
# conda install pytorch torchvision torchaudio pytorch-cuda=12.1 -c pytorch -c nvidia
# (PyTorch officiel supporte CUDA 11.8 et 12.1, pas 13.1 - mais compatible avec driver 591.59)

# Deep Learning & Transformers
# torch>=2.1.0  # ← NE PAS installer via pip, utiliser conda (voir ci-dessus)
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0  # Quantisation 4-bit essentielle pour 4GB VRAM
peft>=0.7.0
sentencepiece>=0.1.99
protobuf>=4.25.0
safetensors>=0.4.1

# Vision & Image Processing
Pillow>=10.1.0
opencv-python-headless>=4.8.1.78
imutils>=0.5.4
scikit-image>=0.22.0
matplotlib>=3.8.0  # Pour visualisations (utilisé dans notebook)

# RAG & Embeddings
langchain>=0.1.0
langchain-community>=0.0.13
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Data Processing
datasets>=2.16.0
pandas>=2.1.4
numpy>=1.24.0,<2.0.0  # Limiter pour compatibilité
scikit-learn>=1.3.2

# Color Processing
webcolors>=1.13

# Web Interface
gradio>=4.13.0
gradio-client>=0.8.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Utilities
tqdm>=4.66.1
huggingface-hub>=0.20.0
requests>=2.31.0

# Optional: Performance monitoring
psutil>=5.9.0  # Monitoring RAM/GPU
gputil>=1.4.0  # Monitoring GPU usage
